+++ src/features/Gw2/actions.js
@@ -61,7 +61,7 @@ export function generateActions (resourceName, getResource, afterGet) {
         return afterGet ? afterGet(dispatch, response) : response;
       }, ({ response }) => {
         const action = response.status === 404
-          ? actions[fetchErrorMethodName](missingIds, T.translate('messages.notFound'))
+          ? actions[fetchErrorMethodName](missingIds, T.translate('messages.notFoundLong'))
           : actions[fetchErrorMethodName](missingIds, T.translate('messages.gw2ApiDown'));
 
         dispatch(action);